Add security enhancements

+++ source/scripts/Scheduler.sh
@@ -1,22 +1,34 @@
-#!/bin/bash -xe
-
-source /etc/environment
-source /root/config.cfg
+#!/bin/bash
 
 if [ $# -lt 2 ]
   then
+    echo "usage: $0 EFS_DATA EFS_APPS"
     exit 1
 fi
 
-# Install SSM
-yum install -y https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
-systemctl enable amazon-ssm-agent
-systemctl restart amazon-ssm-agent
+set -ex
 
-
-mkdir -p /apps/soca/$SOCA_CONFIGURATION
 EFS_DATA=$1
 EFS_APPS=$2
+
+source /etc/environment
+source /root/config.cfg
+source /etc/profile.d/proxy.sh
+
+# Signal Cloudformation if the instance creates successfully or not
+yum -y install -y aws-cfn-bootstrap
+function on_exit {
+    /opt/aws/bin/cfn-signal -e $? --stack ${SOCA_CLOUDFORMATION_STACK} --resource SchedulerEC2Host --region ${AWS_DEFAULT_REGION} || true
+}
+trap on_exit EXIT
+
+# Install SSM
+if ! yum list amazon-ssm-agent; then
+    yum install -y https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+fi
+systemctl enable amazon-ssm-agent || true
+systemctl restart amazon-ssm-agent
+
 SERVER_IP=$(hostname -I)
 SERVER_HOSTNAME=$(hostname)
 SERVER_HOSTNAME_ALT=$(echo $SERVER_HOSTNAME | cut -d. -f1)
@@ -24,22 +36,25 @@ echo $SERVER_IP $SERVER_HOSTNAME $SERVER_HOSTNAME_ALT >> /etc/hosts
 
 if [[ $SOCA_BASE_OS == "rhel7" ]]
 then
-    yum install -y $(echo ${SYSTEM_PKGS[*]}) --enablerepo rhui-REGION-rhel-server-optional
-    yum install -y $(echo ${SCHEDULER_PKGS[*]}) --enablerepo rhui-REGION-rhel-server-optional
+    yum install -y $(echo ${SYSTEM_PKGS[*]} ${SCHEDULER_PKGS[*]}) --enablerepo rhui-REGION-rhel-server-optional
 else
-    yum install -y $(echo ${SYSTEM_PKGS[*]})
-    yum install -y $(echo ${SCHEDULER_PKGS[*]})
+    yum install -y $(echo ${SYSTEM_PKGS[*]} ${SCHEDULER_PKGS[*]})
 fi
 
-yum install -y $(echo ${OPENLDAP_SERVER_PKGS[*]})
-yum install -y $(echo ${SSSD_PKGS[*]})
+yum install -y $(echo ${OPENLDAP_SERVER_PKGS[*]} ${SSSD_PKGS[*]})
 
 # Mount EFS
-mkdir /apps
-mkdir /data
-echo "$EFS_DATA:/ /data/ nfs4 nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport 0 0" >> /etc/fstab
-echo "$EFS_APPS:/ /apps nfs4 nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport 0 0" >> /etc/fstab
+if [ ! -d /data ]; then
+    mkdir -p /data
+    echo "$EFS_DATA:/ /data/ nfs4 nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport 0 0" >> /etc/fstab
+fi
+if [ ! -d /apps ]; then
+    mkdir -p /apps
+    echo "$EFS_APPS:/ /apps nfs4 nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport 0 0" >> /etc/fstab
+fi
 mount -a
+
+mkdir -p /apps/soca/$SOCA_CONFIGURATION
 
 # Install Python if needed
 PYTHON_INSTALLED_VERS=$(/apps/soca/$SOCA_CONFIGURATION/python/latest/bin/python3 --version | awk {'print $NF'})
@@ -63,20 +78,20 @@ else
     echo "Python already installed and at correct version."
 fi
 
-# Install PBSPro if needed
+# Install OpenPBS if needed
 cd ~
-PBSPRO_INSTALLED_VERS=$(/opt/pbs/bin/qstat --version | awk {'print $NF'})
-if [[ "$PBSPRO_INSTALLED_VERS" != "$PBSPRO_VERSION" ]]
+OPENPBS_INSTALLED_VERS=$(/opt/pbs/bin/qstat --version | awk {'print $NF'})
+if [[ "$OPENPBS_INSTALLED_VERS" != "$OPENPBS_VERSION" ]]
 then
-    echo "PBSPro Not Detected, Installing PBSPro ..."
+    echo "OpenPBS Not Detected, Installing OpenPBS ..."
     cd ~
-    wget $PBSPRO_URL
-    if [[ $(md5sum $PBSPRO_TGZ | awk '{print $1}') != $PBSPRO_HASH ]];  then
-        echo -e "FATAL ERROR: Checksum for PBSPro failed. File may be compromised." > /etc/motd
+    wget $OPENPBS_URL
+    if [[ $(md5sum $OPENPBS_TGZ | awk '{print $1}') != $OPENPBS_HASH ]];  then
+        echo -e "FATAL ERROR: Checksum for OpenPBS failed. File may be compromised." > /etc/motd
         exit 1
     fi
-    tar zxvf $PBSPRO_TGZ
-    cd pbspro-$PBSPRO_VERSION
+    tar zxvf $OPENPBS_TGZ
+    cd openpbs-$OPENPBS_VERSION
     ./autogen.sh
     ./configure --prefix=/opt/pbs
     make -j6
@@ -84,7 +99,7 @@ then
     /opt/pbs/libexec/pbs_postinstall
     chmod 4755 /opt/pbs/sbin/pbs_iff /opt/pbs/sbin/pbs_rcp
 else
-    echo "PBSPro already installed, and at correct version."
+    echo "OpenPBS already installed, and at correct version."
     echo "PBS_SERVER=$SERVER_HOSTNAME_ALT
 PBS_START_SERVER=1
 PBS_START_SCHED=1
@@ -137,71 +152,78 @@ EOF
 systemctl enable pbs
 systemctl start pbs
 
-# Default Server config
-/opt/pbs/bin/qmgr -c "create node $SERVER_HOSTNAME_ALT"
-/opt/pbs/bin/qmgr -c "set node $SERVER_HOSTNAME_ALT queue = workq"
-/opt/pbs/bin/qmgr -c "set server flatuid=true"
-/opt/pbs/bin/qmgr -c "set server job_history_enable=1"
-/opt/pbs/bin/qmgr -c "set server job_history_duration = 00:01:00"
-/opt/pbs/bin/qmgr -c "set server scheduler_iteration = 30"
-/opt/pbs/bin/qmgr -c "set server max_concurrent_provision = 5000"
+if [ ! -e /var/lib/cloud/instance/sem/pbs_server_config_done ]; then
+    # Default Server config
+    /opt/pbs/bin/qmgr -c "create node $SERVER_HOSTNAME_ALT"
+    /opt/pbs/bin/qmgr -c "set node $SERVER_HOSTNAME_ALT queue = workq"
+    /opt/pbs/bin/qmgr -c "set server flatuid=true"
+    /opt/pbs/bin/qmgr -c "set server job_history_enable=1"
+    /opt/pbs/bin/qmgr -c "set server job_history_duration = 01:00:00"
+    /opt/pbs/bin/qmgr -c "set server scheduler_iteration = 30"
+    /opt/pbs/bin/qmgr -c "set server max_concurrent_provision = 5000"
+    touch /var/lib/cloud/instance/sem/pbs_server_config_done
+fi
 
-# Default Queue Config
-/opt/pbs/bin/qmgr -c "create queue low"
-/opt/pbs/bin/qmgr -c "set queue low queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue low started = True"
-/opt/pbs/bin/qmgr -c "set queue low enabled = True"
-/opt/pbs/bin/qmgr -c "set queue low default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue normal"
-/opt/pbs/bin/qmgr -c "set queue normal queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue normal started = True"
-/opt/pbs/bin/qmgr -c "set queue normal enabled = True"
-/opt/pbs/bin/qmgr -c "set queue normal default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue high"
-/opt/pbs/bin/qmgr -c "set queue high queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue high started = True"
-/opt/pbs/bin/qmgr -c "set queue high enabled = True"
-/opt/pbs/bin/qmgr -c "set queue high default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue desktop"
-/opt/pbs/bin/qmgr -c "set queue desktop queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue desktop started = True"
-/opt/pbs/bin/qmgr -c "set queue desktop enabled = True"
-/opt/pbs/bin/qmgr -c "set queue desktop default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue test"
-/opt/pbs/bin/qmgr -c "set queue test queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue test started = True"
-/opt/pbs/bin/qmgr -c "set queue test enabled = True"
-/opt/pbs/bin/qmgr -c "set queue test default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue alwayson"
-/opt/pbs/bin/qmgr -c "set queue alwayson queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue alwayson started = True"
-/opt/pbs/bin/qmgr -c "set queue alwayson enabled = True"
-/opt/pbs/bin/qmgr -c  "set server default_queue = normal"
+if [ ! -e /var/lib/cloud/instance/sem/pbs_queue_config_done ]; then
+    # Default Queue Config
+    /opt/pbs/bin/qmgr -c "create queue low"
+    /opt/pbs/bin/qmgr -c "set queue low queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue low started = True"
+    /opt/pbs/bin/qmgr -c "set queue low enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue low default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue normal"
+    /opt/pbs/bin/qmgr -c "set queue normal queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue normal started = True"
+    /opt/pbs/bin/qmgr -c "set queue normal enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue normal default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue high"
+    /opt/pbs/bin/qmgr -c "set queue high queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue high started = True"
+    /opt/pbs/bin/qmgr -c "set queue high enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue high default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue desktop"
+    /opt/pbs/bin/qmgr -c "set queue desktop queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue desktop started = True"
+    /opt/pbs/bin/qmgr -c "set queue desktop enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue desktop default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue test"
+    /opt/pbs/bin/qmgr -c "set queue test queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue test started = True"
+    /opt/pbs/bin/qmgr -c "set queue test enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue test default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue alwayson"
+    /opt/pbs/bin/qmgr -c "set queue alwayson queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue alwayson started = True"
+    /opt/pbs/bin/qmgr -c "set queue alwayson enabled = True"
+    /opt/pbs/bin/qmgr -c "set server default_queue = normal"
 
-# Add compute_node to list of required resource
-sed -i 's/resources: "ncpus, mem, arch, host, vnode, aoe, eoe"/resources: "ncpus, mem, arch, host, vnode, aoe, eoe, compute_node"/g' /var/spool/pbs/sched_priv/sched_config
+    # Add compute_node to list of required resource
+    sed -i 's/resources: "ncpus, mem, arch, host, vnode, aoe, eoe"/resources: "ncpus, mem, arch, host, vnode, aoe, eoe, compute_node"/g' /var/spool/pbs/sched_priv/sched_config
+    touch /var/lib/cloud/instance/sem/pbs_queue_config_done
+fi
 
-# Configure Ldap
-systemctl enable slapd
-systemctl start slapd
+if [ ! -e /var/lib/cloud/instance/sem/ldap_configured ]; then
+    # Configure Ldap
+    systemctl enable slapd
+    systemctl start slapd
 
-MASTER_LDAP_PASSWORD=$(slappasswd -g)
-MASTER_LDAP_PASSWORD_ENCRYPTED=$(/sbin/slappasswd -s $MASTER_LDAP_PASSWORD -h "{SSHA}")
-echo -n "admin" > /root/OpenLdapAdminUsername.txt
-echo -n $MASTER_LDAP_PASSWORD > /root/OpenLdapAdminPassword.txt
-chmod 600 /root/OpenLdapAdminPassword.txt
-echo "URI ldap://$SERVER_HOSTNAME" >> /etc/openldap/ldap.conf
-echo "BASE $LDAP_BASE" >> /etc/openldap/ldap.conf
+    MASTER_LDAP_PASSWORD=$(slappasswd -g)
+    MASTER_LDAP_PASSWORD_ENCRYPTED=$(/sbin/slappasswd -s $MASTER_LDAP_PASSWORD -h "{SSHA}")
+    echo -n "admin" > /root/OpenLdapAdminUsername.txt
+    echo -n $MASTER_LDAP_PASSWORD > /root/OpenLdapAdminPassword.txt
+    chmod 600 /root/OpenLdapAdminPassword.txt
+    echo "URI ldap://$SERVER_HOSTNAME" >> /etc/openldap/ldap.conf
+    echo "BASE $LDAP_BASE" >> /etc/openldap/ldap.conf
 
-# Generate 10y certificate for ldaps
-openssl req -new -newkey rsa:2048 -days 3650 -nodes -x509 \
-    -subj "/C=US/ST=California/L=Sunnyvale/O=Aligo/CN=$SERVER_HOSTNAME" \
-    -keyout /etc/openldap/certs/soca.key -out /etc/openldap/certs/soca.crt
+    # Generate 10y certificate for ldaps
+    openssl req -new -newkey rsa:2048 -days 3650 -nodes -x509 \
+        -subj "/C=US/ST=California/L=Sunnyvale/O=Aligo/CN=$SERVER_HOSTNAME" \
+        -keyout /etc/openldap/certs/soca.key -out /etc/openldap/certs/soca.crt
 
-chown ldap:ldap /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
-chmod 600 /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
+    chown ldap:ldap /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
+    chmod 600 /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
 
-echo -e "
+    echo -e "
 dn: olcDatabase={2}hdb,cn=config
 changetype: modify
 replace: olcSuffix
@@ -218,7 +240,7 @@ replace: olcRootPW
 olcRootPW: $MASTER_LDAP_PASSWORD_ENCRYPTED
 " > db.ldif
 
-echo -e "
+    echo -e "
 dn: cn=config
 changetype: modify
 replace: olcTLSCertificateFile
@@ -228,7 +250,7 @@ replace: olcTLSCertificateKeyFile
 olcTLSCertificateKeyFile: /etc/openldap/certs/soca.key
 " > update_ssl_cert.ldif
 
-echo -e "
+    echo -e "
 dn: olcDatabase={2}hdb,cn=config
 changetype: modify
 replace: olcAccess
@@ -238,7 +260,7 @@ add: olcAccess
 olcAccess: {1}to * by dn.base="gidNumber=0+uidNumber=0,cn=peercred,cn=external,cn=auth" write by dn.base="ou=admins,$LDAP_BASE" write by * read
 " > change_user_password.ldif
 
-echo -e "
+    echo -e "
 dn: cn=sudo,cn=schema,cn=config
 objectClass: olcSchemaConfig
 cn: sudo
@@ -252,15 +274,15 @@ olcAttributeTypes: ( 1.3.6.1.4.1.15953.9.1.7 NAME 'sudoRunAsGroup' DESC 'Group(s
 olcObjectClasses: ( 1.3.6.1.4.1.15953.9.2.1 NAME 'sudoRole' SUP top STRUCTURAL DESC 'Sudoer Entries' MUST ( cn ) MAY ( sudoUser $ sudoHost $ sudoCommand $ sudoRunAs $ sudoRunAsUser $ sudoRunAsGroup $ sudoOption $ description ) )
 " > sudoers.ldif
 
-/bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f db.ldif
-/bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f update_ssl_cert.ldif
-/bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f change_user_password.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f sudoers.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/cosine.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/nis.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/inetorgperson.ldif
+    /bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f db.ldif
+    /bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f update_ssl_cert.ldif
+    /bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f change_user_password.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f sudoers.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/cosine.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/nis.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/inetorgperson.ldif
 
-echo -e "
+    echo -e "
 dn: $LDAP_BASE
 dc: soca
 objectClass: top
@@ -287,24 +309,24 @@ objectClass: organizationalUnit
 ou: Group
 " > base.ldif
 
-/bin/ldapadd -x -W -y /root/OpenLdapAdminPassword.txt -D "cn=admin,$LDAP_BASE" -f base.ldif
+    /bin/ldapadd -x -W -y /root/OpenLdapAdminPassword.txt -D "cn=admin,$LDAP_BASE" -f base.ldif
 
-authconfig \
-    --enablesssd \
-    --enablesssdauth \
-    --enableldap \
-    --enableldapauth \
-    --ldapserver="ldap://$SERVER_HOSTNAME" \
-    --ldapbasedn="$LDAP_BASE" \
-    --enablelocauthorize \
-    --enablemkhomedir \
-    --enablecachecreds \
-    --updateall
+    authconfig \
+        --enablesssd \
+        --enablesssdauth \
+        --enableldap \
+        --enableldapauth \
+        --ldapserver="ldap://$SERVER_HOSTNAME" \
+        --ldapbasedn="$LDAP_BASE" \
+        --enablelocauthorize \
+        --enablemkhomedir \
+        --enablecachecreds \
+        --updateall
 
-echo "sudoers: files sss" >> /etc/nsswitch.conf
+    echo "sudoers: files sss" >> /etc/nsswitch.conf
 
-# Configure SSSD
-echo -e "[domain/default]
+    # Configure SSSD
+    echo -e "[domain/default]
 enumerate = True
 autofs_provider = ldap
 cache_credentials = True
@@ -343,10 +365,13 @@ ldap_sudo_smart_refresh_interval=3600
 [ifp]
 
 [secrets]" > /etc/sssd/sssd.conf
-chmod 600 /etc/sssd/sssd.conf
+    chmod 600 /etc/sssd/sssd.conf
 
-systemctl enable sssd
-systemctl restart sssd
+    systemctl enable sssd
+    systemctl restart sssd
+
+    touch /var/lib/cloud/instance/sem/ldap_configured
+fi
 
 # Disable SELINUX
 sed -i 's/SELINUX=enforcing/SELINUX=disabled/g' /etc/selinux/config
@@ -359,15 +384,11 @@ echo "UserKnownHostsFile /dev/null" >> /etc/ssh/ssh_config
 # Source environment to reload path for Python3
 /apps/soca/$SOCA_CONFIGURATION/python/$PYTHON_VERSION/bin/pip3 install -r /root/requirements.txt
 
-# PBS hooks still run on python2 env, so some packages are required
-EASY_INSTALL=$(which easy_install-2.7)
-$EASY_INSTALL yaml python-ldap boto3
-
-# Configure NTP
-yum remove -y ntp
-yum install -y chrony
-mv /etc/chrony.conf  /etc/chrony.conf.original
-echo -e """
+if [ ! -e /var/lib/cloud/instance/sem/chrony_configured ]; then
+    # Configure Chrony
+    yum remove -y ntp
+    mv /etc/chrony.conf  /etc/chrony.conf.original
+    echo -e """
 # use the local instance NTP service, if available
 server 169.254.169.123 prefer iburst minpoll 4 maxpoll 4
 
@@ -394,14 +415,23 @@ logdir /var/log/chrony
 dumponexit
 dumpdir /var/run/chrony
 """ > /etc/chrony.conf
-systemctl enable chronyd
+    systemctl enable chronyd
+    touch /var/lib/cloud/instance/sem/chrony_configured
+fi
 
 # Disable ulimit
 echo -e  "
 * hard memlock unlimited
 * soft memlock unlimited
-" > /etc/security/limits.conf
+" >> /etc/security/limits.conf
 
 # Reboot to ensure SELINUX is disabled
 # Note: Upon reboot, SchedulerPostReboot.sh script will be executed and will finalize scheduler configuration
-reboot
+if [ ! -e /var/lib/cloud/instance/sem/rebooted ] || ! needs-restarting -r; then
+    touch /var/lib/cloud/instance/sem/rebooted
+    reboot
+else
+    aws s3 cp s3://${SOCA_INSTALL_BUCKET}/${SOCA_INSTALL_BUCKET_FOLDER}/scripts/SchedulerPostReboot.sh /root/
+    chmod +x /root/SchedulerPostReboot.sh
+    /root/call-SchedulerPostReboot.sh >> /root/PostRebootConfig.log 2>&1
+fi
